Moved all non-critical code to a scrap folder for now, going to gradually bring it back in and refactor it.
